+++ packages/react-native-gradle-plugin/src/main/kotlin/com/facebook/react/utils/NdkConfiguratorUtils.kt
@@ -103,6 +103,7 @@ internal object NdkConfiguratorUtils {
               "**/libreact_nativemodule_core.so",
               "**/libreact_newarchdefaults.so",
               "**/libreact_cxxreactpackage.so",
+              "**/libreact_defaults.so",
               "**/libreact_render_componentregistry.so",
               "**/libreact_render_core.so",
               "**/libreact_render_debug.so",
