+++ packages/react-native/ReactAndroid/cmake-utils/ReactNative-application.cmake
@@ -71,6 +71,7 @@ add_library(react_utils ALIAS ReactAndroid::react_utils)
 add_library(react_render_componentregistry ALIAS ReactAndroid::react_render_componentregistry)
 add_library(react_newarchdefaults ALIAS ReactAndroid::react_newarchdefaults)
 add_library(react_cxxreactpackage ALIAS ReactAndroid::react_cxxreactpackage)
+add_library(react_defaults ALIAS ReactAndroid::react_defaults)
 add_library(react_render_core ALIAS ReactAndroid::react_render_core)
 add_library(react_render_graphics ALIAS ReactAndroid::react_render_graphics)
 add_library(rrc_view ALIAS ReactAndroid::rrc_view)
@@ -100,6 +101,7 @@ target_link_libraries(${CMAKE_PROJECT_NAME}
         react_nativemodule_core             # prefab ready
         react_newarchdefaults               # prefab ready
         react_cxxreactpackage               # prefab ready
+        react_defaults                      # prefab ready
         react_render_componentregistry      # prefab ready
         react_render_core                   # prefab ready
         react_render_debug                  # prefab ready
